requirements.txt
python-dotenv
azure-ai-vision
opencv-python
numpy
flask